% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_functions.R
\name{nuc_number}
\alias{nuc_number}
\title{Nucleotide Number Function}
\usage{
nuc_number(rawDNA)
}
\arguments{
\item{rawDNA}{This parameter is input of the raw DNA string or a variable/file containing the DNA string.}
}
\description{
This function counts the number of nucleotides in the DNA string.
}
\examples{
nuc_number()
}
\keyword{DNA,}
\keyword{counting}
\keyword{nucleotide,}
\keyword{number,}
