% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_functions.R
\name{transcription}
\alias{transcription}
\title{Transcription Function}
\usage{
transcription(rawDNA)
}
\arguments{
\item{rawDNA}{This parameter is input of the raw DNA string or a variable/file containing the DNA string.}
}
\description{
This function converts a DNA string to the corresponding RNA string.
}
\examples{
transcription()
}
\keyword{DNA,}
\keyword{RNA,}
\keyword{Transcription}
